Remove legacy language machinery
