+++ packages/cursorless-engine/src/languages/TreeSitterQuery/makeRangeFromPositions.ts
@@ -1,0 +1,14 @@
+import { Range } from "@cursorless/common";
+import type { Point } from "web-tree-sitter";
+
+export function makeRangeFromPositions(
+  startPosition: Point,
+  endPosition: Point,
+) {
+  return new Range(
+    startPosition.row,
+    startPosition.column,
+    endPosition.row,
+    endPosition.column,
+  );
+}
