+++ packages/cursorless-engine/src/processTargets/ModifierStageFactory.ts
@@ -7,7 +7,4 @@ import type { ModifierStage } from "./PipelineStages.types";
 
 export interface ModifierStageFactory {
   create(modifier: Modifier): ModifierStage;
-  getLegacyScopeStage(
-    modifier: ContainingScopeModifier | EveryScopeModifier,
-  ): ModifierStage;
 }
