+++ packages/cursorless-engine/src/processTargets/ModifierStageFactoryImpl.ts
@@ -30,11 +30,6 @@ import { RawSelectionStage } from "./modifiers/RawSelectionStage";
 import { RelativeScopeStage } from "./modifiers/RelativeScopeStage";
 import { VisibleStage } from "./modifiers/VisibleStage";
 import type { ScopeHandlerFactory } from "./modifiers/scopeHandlers/ScopeHandlerFactory";
-import type {
-  SimpleContainingScopeModifier,
-  SimpleEveryScopeModifier,
-} from "./modifiers/scopeTypeStages/LegacyContainingSyntaxScopeStage";
-import { LegacyContainingSyntaxScopeStage } from "./modifiers/scopeTypeStages/LegacyContainingSyntaxScopeStage";
 
 export class ModifierStageFactoryImpl implements ModifierStageFactory {
   constructor(
@@ -114,32 +109,6 @@ export class ModifierStageFactoryImpl implements ModifierStageFactory {
         const { type } = modifier;
         throw new Error(`Unknown modifier: ${type}`);
       }
-    }
-  }
-
-  /**
-   * Any scope type that has not been fully migrated to the new
-   * {@link ScopeHandler} setup should have a branch in this `switch` statement.
-   * Once the scope type is fully migrated, remove the branch and the legacy
-   * modifier stage.
-   *
-   * Note that it is possible for a scope type to be partially migrated.  For
-   * example, we could support modern scope handlers for a certain scope type in
-   * Ruby, but not yet in Python.
-   *
-   * @param modifier The modifier for which to get the modifier stage
-   * @returns A scope stage implementing the modifier for the given scope type
-   */
-  getLegacyScopeStage(
-    modifier: ContainingScopeModifier | EveryScopeModifier,
-  ): ModifierStage {
-    switch (modifier.scopeType.type) {
-      default:
-        // Default to containing syntax scope using tree sitter
-        return new LegacyContainingSyntaxScopeStage(
-          this.languageDefinitions,
-          modifier as SimpleContainingScopeModifier | SimpleEveryScopeModifier,
-        );
     }
   }
 }
