+++ packages/cursorless-engine/src/processTargets/modifiers/EveryScopeStage.ts
@@ -48,9 +48,7 @@ export class EveryScopeStage implements ModifierStage {
     );
 
     if (scopeHandler == null) {
-      return this.modifierStageFactory
-        .getLegacyScopeStage(this.modifier)
-        .run(target, options);
+      throw new NoContainingScopeError(scopeType.type);
     }
 
     let scopes: TargetScope[] | undefined;
