--- packages/cursorless-vscode-e2e/src/suite/cursorlessEngineIntegration.vscode.test.ts
@@ -1,12 +1,0 @@
-import { getCursorlessApi } from "@cursorless/vscode-common";
-import { endToEndTestSetup } from "../endToEndTestSetup";
-
-suite("Cursorless engine integration", async function () {
-  endToEndTestSetup(this);
-
-  test("integration", async () => {
-    const cursorlessApi = await getCursorlessApi();
-    const { runIntegrationTests } = cursorlessApi.testHelpers!;
-    await runIntegrationTests();
-  });
-});
